+++ resolwe/flow/utils/purge.py
@@ -4,24 +4,17 @@ Data Purge
 ==========
 
 """
+from __future__ import absolute_import, division, print_function, unicode_literals
+
 import os
 import shutil
 
 from django.conf import settings
 
-from flow.models import Data, iterate_fields, Storage
+from resolwe.flow.models import Data, Process, iterate_fields, Storage
 
 
-def data_purge(data_ids=None, delete=False, verbosity=1):
-    """Print files not referenced from meta data.
-
-    If data_ids not given, run on all data objects.
-    If delete is True, delete unreferenced files.
-
-    """
-    data_dir = settings.FLOW_EXECUTOR['DATA_DIR']
-    path_offset = len(data_dir) + 24 + 2
-
+def get_purge_files(root, output, output_schema, descriptor, descriptor_schema):
     def remove_file(fn, paths):
         """From paths remove fn and dirs before fn in dir tree."""
         while fn:
@@ -46,107 +39,98 @@ def data_purge(data_ids=None, delete=False, verbosity=1):
         """Extend unreferenced list with all subdirs and files in top dir."""
         subs = []
         for path, dirs, files in os.walk(root, topdown=False):
-            path = path[path_offset:]
+            path = path[len(root) + 1:]
             subs.extend(os.path.join(path, f) for f in files)
             subs.extend(os.path.join(path, d) for d in dirs)
         return subs
 
-    check = Data.objects.all()
-    purge = []
-    ref_storage = set()
-    finished = [Data.STATUS_DONE, Data.STATUS_ERROR]
-    if data_ids:
-        check = check.filter(id__in=data_ids)
+    unreferenced_files = subfiles(root)
 
-        purge = set(data_ids)
-    else:
-        purge = set(os.listdir(data_dir))
+    remove_file('jsonout.txt', unreferenced_files)
+    remove_file('stderr.txt', unreferenced_files)
+    remove_file('stdout.txt', unreferenced_files)
 
-    # Remove not referenced
-    purge = purge.difference(str(d.id) for d in check.only('pk'))
-    purge = [os.path.join(data_dir, p) for p in purge]
+    meta_fields = [
+        [output, output_schema],
+        [descriptor, descriptor_schema]
+    ]
 
-    check = check.filter(status__in=finished)
+    for meta_field, meta_field_schema in meta_fields:
+        for field_schema, fields in iterate_fields(meta_field, meta_field_schema):
+            if 'type' in field_schema:
+                field_type = field_schema['type']
+                field_name = field_schema['name']
 
-    for d in check:
-        root = os.path.join(data_dir, str(d.id))
-        subs = subfiles(root)
+                # Remove basic:file: entries
+                if field_type.startswith('basic:file:'):
+                    remove_file(fields[field_name]['file'], unreferenced_files)
 
-        meta_fields = [
-            ['output', 'output_schema'],
-            ['static', 'static_schema'],
-            ['var', 'var_template'],
-        ]
+                # Remove list:basic:file: entries
+                elif field_type.startswith('list:basic:file:'):
+                    for field in fields[field_name]:
+                        remove_file(field['file'], unreferenced_files)
 
-        for meta_field, meta_field_schema in meta_fields:
-            for field_schema, fields in iterate_fields(d[meta_field], d[meta_field_schema]):
-                if 'type' in field_schema:
-                    field_type = field_schema['type']
-                    field_name = field_schema['name']
+                # Remove basic:dir: entries
+                elif field_type.startswith('basic:dir:'):
+                    remove_tree(fields[field_name]['dir'], unreferenced_files)
 
-                    # Remove basic:file: entries
-                    if field_type.startswith('basic:file:'):
-                        remove_file(fields[field_name]['file'], subs)
+                # Remove list:basic:dir: entries
+                elif field_type.startswith('list:basic:dir:'):
+                    for field in fields[field_name]:
+                        remove_tree(field['dir'], unreferenced_files)
 
-                    # Remove list:basic:file: entries
-                    elif field_type.startswith('list:basic:file:'):
-                        for field in fields[field_name]:
-                            remove_file(field['file'], subs)
+                # Remove refs entries
+                if field_type.startswith('basic:file:') or field_type.startswith('basic:dir:'):
+                    for ref in fields[field_name].get('refs', []):
+                        remove_tree(ref, unreferenced_files)
 
-                    # Remove basic:url: entries
-                    elif field_type.startswith('basic:url:'):
-                        remove_file(fields[field_name]['url'], subs)
+                elif field_type.startswith('list:basic:file:') or field_type.startswith('list:basic:dir:'):
+                    for field in fields[field_name]:
+                        for ref in field.get('refs', []):
+                            remove_tree(ref, unreferenced_files)
 
-                    # Remove list:basic:url: entries
-                    elif field_type.startswith('list:basic:url:'):
-                        for field in fields[field_name]:
-                            remove_file(field['url'], subs)
+    return set([os.path.join(root, filename) for filename in unreferenced_files])
 
-                    # Remove refs entries
-                    if field_type.startswith('basic:file:') or field_type.startswith('basic:url:'):
-                        for ref in fields[field_name].get('refs', []):
-                            remove_tree(ref, subs)
 
-                    elif field_type.startswith('list:basic:file:') or field_type.startswith('list:basic:url:'):
-                        for field in fields[field_name]:
-                            for ref in field.get(['refs'], []):
-                                remove_tree(ref, subs)
+def data_purge(data_ids=None, delete=False, verbosity=0):
+    """Print files not referenced from meta data.
 
-                    if field_schema['type'].startswith('basic:json:'):
-                        ref_storage.add(fields[field_name])
+    If data_ids not given, run on all data objects.
+    If delete is True, delete unreferenced files.
 
-        purge.extend(os.path.join(root, f) for f in subs)
+    """
 
-    # IMPORTANT: Unreferenced `Storage` objects can be determined only
-    #            if all `Data` objects are checked.
-    unref_storage = Storage.objects.none()
-    if (verbosity or delete) and data_ids is None:
-        unref_storage = Storage.objects.filter(pk__nin=ref_storage).only('pk')
+    data_path = settings.FLOW_EXECUTOR['DATA_DIR']
+    unreferenced_files = set()
+
+    data_qs = Data.objects.filter(status__in=[Data.STATUS_DONE, Data.STATUS_ERROR])
+    if data_ids is not None:
+        data_qs = data_qs.filter(pk__in=data_ids)
+
+    for data in data_qs:
+        root = os.path.join(data_path, str(data.id))
+
+        unreferenced_files.update(get_purge_files(
+            root,
+            data.output,
+            data.process.output_schema,
+            data.descriptor,
+            getattr(data.descriptor_schema, 'schema', [])
+        ))
 
     if verbosity >= 1:
         # Print unreferenced files
-        if purge:
-            print("Unreferenced files ({}):".format(len(purge)))
-            for name in purge:
+        if unreferenced_files:
+            print("Unreferenced files ({}):".format(len(unreferenced_files)))
+            for name in unreferenced_files:
                 print("  {}".format(name))
         else:
             print("No unreferenced files")
 
-        if unref_storage:
-            print("Unreferenced Storage objects ({}):".format(unref_storage.count()))
-            for s in unref_storage:
-                print("  {}".format(s.pk))
-        elif data_ids is None:
-            print("No unreferenced Storage objects")
-
     # Go through unreferenced files and delete them
     if delete:
-        for name in purge:
+        for name in unreferenced_files:
             if os.path.isfile(name) or os.path.islink(name):
                 os.remove(name)
             elif os.path.isdir(name):
                 shutil.rmtree(name)
-
-        # Safety check to prevent deleting all `Storage` objects
-        if data_ids is None:
-            unref_storage.delete()