+++ docs/CHANGELOG.rst
@@ -23,6 +23,7 @@ Added
 - ``RESOLWE_CUSTOM_TOOLS_PATHS`` setting to support custom paths for tools
   directories
 - Add test coverage and track it with Codecov
+- Implement data purge
 
 Changed
 -------